+++ docs/release-notes.md
@@ -1,5 +1,8 @@
 # Release notes
 
+### [x.x.x]
+- Improve the error message when some CI environment variables are missing.
+
 ### [1.17.4]
 - Fix order of checking if a package version exists and forcing package versions